+++ app/build.gradle
@@ -55,7 +55,6 @@ dependencies {
     implementation 'com.fasterxml.jackson.core:jackson-databind:2.9.8'
     implementation 'androidx.recyclerview:recyclerview:1.1.0'
     implementation 'devs.mulham.horizontalcalendar:horizontalcalendar:1.3.4'
-    implementation 'com.github.TutorialsAndroid:GButton:v1.0.19'
     implementation "androidx.room:room-runtime:$room_version"
     annotationProcessor "androidx.room:room-compiler:$room_version"
     testImplementation "androidx.room:room-testing:$room_version"
