+++ app/src/main/AndroidManifest.xml
@@ -18,7 +18,6 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".activities.GoogleSingInActivity"></activity>
         <activity android:name=".activities.AuthActivity" />
         <activity
             android:name=".activities.SettingsActivity"
