+++ app/src/main/res/layout/activity_login.xml
@@ -78,11 +78,9 @@
     </ScrollView>
 
 
-    <com.developer.gbuttons.GoogleSignInButton
-        android:id="@+id/google_sign_in_button"
+    <com.google.android.gms.common.SignInButton
+        android:id="@+id/google_sign_in_button_2"
         android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="@string/google_sign_up"
-        app:isDarkTheme="true" />
+        android:layout_height="wrap_content" />
 
 </LinearLayout>
