+++ app/src/main/res/xml/network_security_config.xml
@@ -1,6 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
+
 <network-security-config>
     <domain-config cleartextTrafficPermitted="true">
-        <domain includeSubdomains="true">192.168.1.119</domain>
+        <domain includeSubdomains="true">192.168.0.104</domain>
     </domain-config>
 </network-security-config>
